requirements.txt
streamlit
Pillow
google-generativeai
python-dotenv
jsonschema
opencv-python
numpy